mednex-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.40.0
torch==2.5.1
pymongo[srv]==4.6.0
dnspython==2.4.2
pandas==2.1.3
networkx==3.2.1
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
scikit-learn==1.3.2
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
pydantic[email]==2.5.0
